+++ mrblib/acli/socket.rb
@@ -1,17 +1,20 @@
 module Acli
   class Socket
-    attr_reader :ws, :tls
+    attr_reader :ws, :tls, :logger
     alias tls? tls
 
-    def initialize(uri, headers = nil, protocol = "actioncable-v1-json")
+    def initialize(uri, headers: nil, protocol: "actioncable-v1-json", logger: NoopLogger.new)
+      @logger = logger
       @tls = uri.scheme == "https"
 
       connection_class = tls? ? WebSocket::WssConnection : WebSocket::WsConnection
 
       fullpath = "#{uri.path}#{uri.query ? "?#{uri.query}" : ""}"
+      logger.log "Socket connection params: host=#{uri.host} port=#{uri.port} path=#{fullpath} tls=#{tls?}"
       @ws = connection_class.new(uri.host, uri.port, fullpath, tls_config)
       ws.custom_headers = headers
       ws.protocol = protocol
+      ws.logger = logger
 
       setup_ws
     end
@@ -32,6 +35,7 @@ module Acli
     end
 
     def send(msg, opcode = nil, timeout = -1)
+      logger.log "Socket send: #{msg}"
       ws.send(msg, opcode, timeout)
     end
 
@@ -39,6 +43,8 @@ module Acli
       return unless frame
 
       if frame.opcode == :connection_close
+        logger.log "Socket received close: #{frame.status_code}"
+
         raise Acli::ClonnectionClosedError, "Closed with status: #{frame.status_code}"
       end
 
@@ -46,6 +52,7 @@ module Acli
     end
 
     def close(status_code = :normal_closure, reason = nil, timeout = -1)
+      logger.log "Socket closed: #{status_code}"
       ws.close(status_code, reason, timeout)
     end